fixed couple mistakes, replaced TesTokenInterceptor to controllers, refactored APIClient class
